Began refactoring.
