+++ core.py
@@ -4,83 +4,14 @@ import schedule
 import requests
 import lxml.html as html
 import unicodedata
-from config import telegram_api_key
 import time
+from datetime import datetime
 
+from common.config import telegram_api_key
+from common.vars import kinochat_id, members_list, stickers_list, genre_list
+from common.helpers import random_member, get_random_actor
 
 bot = telebot.TeleBot(telegram_api_key)
-kinochat_id = -1001323316776
-members_list = [
-    {
-        'username': '@molotoko',
-        'gender': 'female'
-    },
-    {
-        'username': '@HKEY47',
-        'gender': 'male'
-    },
-    {
-        'username': '@mikmall',
-        'gender': 'male'
-    },
-    {
-        'username': '@madurmanov',
-        'gender': 'male'
-    },
-    {
-        'username': '@fyvdlo',
-        'gender': 'male'
-    },
-    {
-        'username': '@wilddeer',
-        'gender': 'male'
-    },
-    {
-        'username': '@Henpukuhime',
-        'gender': 'female'
-    },
-    {
-        'username': '@anechka_persik',
-        'gender': 'female'
-    },
-    {
-        'username': '@sleepercat0_0',
-        'gender': 'female'
-    },
-    {
-        'username': '@Milli_M',
-        'gender': 'female'
-    },
-    {
-        'username': '@Dart_gedark',
-        'gender': 'male'
-    },
-    {
-        'username': '@nogpyra',
-        'gender': 'female'
-    },
-    {
-        'username': '@tsynali',
-        'gender': 'female'
-    },
-    {
-        'username': '@kokos_89',
-        'gender': 'male'
-    },
-    {
-        'username': '@beforescriptum',
-        'gender': 'female'
-    },
-    {
-        'username': '@MoshWayne',
-        'gender': 'male'
-    }
-
-]
-stickers_list = {
-    'droed': 'CAACAgIAAxkBAAO1Xrs0GAK_ts-_2AG5lhTO2VwRTS4AAl0BAAJEyQkHfIbn433Oi2gZBA',
-    'droed_work': 'CAACAgIAAxkBAAOvXrszrgvwIgKSJj105YntGMYTL7cAAl4BAAJEyQkHQhFYn9ziwn4ZBA'
-}
 
 
 @bot.message_handler(commands=['start'])
@@ -142,9 +73,6 @@ def random_film(message):
 
 @bot.message_handler(commands=['genre'])
 def random_genre(message):
-    genre_list = ['аниме', 'биография', 'боевик', 'вестерн', 'военный', 'детектив', 'документальный', 'драма',
-                  'история', 'комедия', 'короткометражка', 'криминал', 'мелодрама', 'музыка', 'мультфильм',
-                  'приключения', 'семейный', 'спорт', 'триллер', 'ужасы', 'фантастика', 'фильм-нуар', 'фэнтези']
     genre_value = random.choice(genre_list)
 
     erotic = ' с эротикой'
@@ -161,12 +89,9 @@ def random_genre(message):
 
 @bot.message_handler(commands=['vip'])
 def get_vip(message):
-    random_member = random.choice(members_list)
-    random_member_username = random_member['username']
-
     bot.send_message(
         message.chat.id,
-        f'Киноман *{random_member_username}* получил статус *VIP*. Это значит, что он имеет 4 голоса вместо 2 и может '
+        f'Киноман *{random_member()}* получил статус *VIP*. Это значит, что он имеет 4 голоса вместо 2 и может '
         f'распределить их как угодно, но не более 2 голосов за 1 фильм.',
         parse_mode='Markdown'
     )
@@ -174,12 +99,9 @@ def get_vip(message):
 
 @bot.message_handler(commands=['kinoman'])
 def kinoman_of_the_week(message):
-    random_member = random.choice(members_list)
-    random_member_username = random_member['username']
-
     bot.send_message(
         message.chat.id,
-        f'Киноман *{random_member_username}* становится почетным *Киноманом Недели*. '
+        f'Киноман *{random_member()}* становится почетным *Киноманом Недели*. '
         f'Он может предложить любой фильм к просмотру вне очереди!',
         parse_mode='Markdown'
     )
@@ -187,12 +109,9 @@ def kinoman_of_the_week(message):
 
 @bot.message_handler(commands=['santa'])
 def santa(message):
-    random_member = random.choice(members_list)
-    random_member_username = random_member['username']
-
     bot.send_message(
         message.chat.id,
-        f'Киноман *{random_member_username}* становится Сантой. '
+        f'Киноман *{random_member()}* становится Сантой. '
         f'Он может предложить любой фильм к просмотру вне очереди!\n'
         f'Этот фильм мы занесем в Коллективные просмотры!',
         parse_mode='Markdown'
@@ -203,32 +122,6 @@ def santa(message):
 def go_to_work(message):
     bot.send_sticker(message.chat.id, stickers_list['droed'])
     bot.send_sticker(message.chat.id, stickers_list['droed_work'])
-
-
-def get_random_actor(gender):
-    link = 'https://www.imdb.com/search/name/?groups=oscar_winner,oscar_nominee&count=100'
-
-    if gender:
-        link = f'{link}&gender={gender}'
-
-    actors_request = requests.get(link)
-    actors_tree = html.fromstring(actors_request.text)
-    actors = actors_tree.find_class('lister-item')
-
-    actor = random.choice(actors).find_class('lister-item-header')[0].find('a')
-    actor_name = actor.text_content().strip()
-    actor_href = actor.attrib['href']
-    actor_link = f'https://www.imdb.com{actor_href}'
-
-    actor_request = requests.get(actor_link)
-    actor_tree = html.fromstring(actor_request.text)
-    actor_death_info = actor_tree.xpath('//div[@id="name-death-info"]')
-
-    return {
-        'name': actor_name,
-        'link': actor_link,
-        'alive': False if len(actor_death_info) else True
-    }
 
 
 @bot.message_handler(commands=['date'])
@@ -290,7 +183,7 @@ def born_today(message):
     )
 
 
-def born_today_cron():
+def born_today_results():
     request_link = 'https://www.imdb.com/feature/bornondate/'
     page = requests.get(request_link)
     tree = html.fromstring(page.text)
@@ -315,11 +208,13 @@ def born_today_cron():
     )
 
 
-def notifications():
-    schedule.every().day.at('10:00').do(born_today_cron())
-    while True:
-        schedule.run_pending()
-        time.sleep(1)
+@bot.message_handler(func=lambda message: False)
+def scheduled_events():
+    now = datetime.now()
+    current_time = now.strftime("%H:%M:%S")
+
+    if current_time == '18:16:00':
+        born_today_results()
 
 
 @bot.message_handler(content_types=['text'])
